very very messy commit with a working prototype of a reactive composable
The graphdata.js composable sets up a globally reactive state which
is imported by the root component App.vue, from which it is in turn
provided to children components. They inject the graph and related
methods and can then use them as necessary. Additionally, the
PropertyShapeEditor component is updated to assign a uuid to each
instance upon 'onBeforeMount' lifecycle hook firing. This also creates
an empty triple in the global graph database. Finally, the editor
component, e.g. UIEditor, injects the graph and gets the uuid as a prop,
and thereby accesses the ccorrect graph entry for which to set the object.
Lots more to be done...