Add script to extract class hierarchy related triples
In order to present users for all the correct possibilities
in a list of options in a form, the list should not only present
all nodes with the type of class expected by the list, but
also all nodes with a class type that is a subclass of the
originating super class. For this reason, all triples that specify
aspects of the class hierarchy of the form schema should be available
to the list of options. These triples aren't part of a shacl export
of a linkml schema, but they are included in an OWL export. This
script allows extracting said triples from an OWL export of a
schema and dumps them in a ttl file for use by the shacl-vue app