Laying the foundation for adding and saving multiple nodes and triples
This commit contains changes across the major components and composables
involved in redering the form and its fields. The goal is to allow an array
of values to be entered for fields that allow this (based on sh:maxCount).
This includes restructuring the internal representation of formData so that
the correct element in an array of nodes or triples is being edited, new
functionality for adding empty nodes/triples, new functionality for removing
previously entered triples.

An important addition, which should be remembered for future use, is the
use of the 'key' property on any v-for component in order to make the
component instances unique. If the key changes, a new component instance
is created. This prevents bugs specifically in the case where a node is saved
and a new empty node should take its place.