Include subclasses and 'add new item' in InstanceSelectEditor
When users want to select an item from a dropdown list with a specific
target class (e.g. Contributor), the list should also show all items
that belong to the subclasses of the target class (e.g. Person or
Organization). In addition, if the preferred item is not found, the
list should allow the option of adding more items, also belonging to
both the target class and all subclasses. This commit provides changes
to achieve this. The selection of a specific class to add as a new item,
as well as entering the fields for the new item, remains as a TODO.