Abstract out the FormEditor component
The relevant functionality is taken out of MainForm and put
into a new component. This allows the formeditor component
to be used in totality as a standalone form, e.g. inside a
dialog that is to be used within the InstanceSelectEditor
when adding a new instance is necessary. The formeditor
component additionally includes the groundwork for tracking
all of its child fields/inputs, so that the form validation
functionality can have access to all fields and their errors