Adds new refregister composable
This allows any editor components to register their template ref
of their input to their (grand)parent FormEditor component. This is necessary
for validation, and allows the matching of an input id (as a DOM element
which the validation errors reference) to a specific ref of an input from
which we can get more informative information (such as field name, label, type,
and whatever else is related/provided to the ref). This essentially allows us
to show a list of informative validation error messages for all inputs on a
form, irrespective of their depth in the component hierarchy.