New base input composable to abstract common editor functionality
This commit brings an important change to the application in terms
of abstracting common editor functionality into a composable. While
some editors consist of a single base component, such a text field
or a select, it could be quite common that an editor component might
have a set of interdependent components with (possibly complex)
interdependent logic. Ideally, these complexities should be contained
within the editor component itself, and should not be exposed to the
parent or rest of the application. Similarly, the effort for a developer
to create their own custom editor component should ideally be kept low
and they shouldn't have to use all of the app-specific inner logic.
Therefore, it made sense to:
- create a simplified top-down API for new custom editors
- provide 'base class' functionality that new custom editors can import

The simplified API basically means that custom editors do not have to
receive, understand, or manipulate the form data structure anymore in
order to achieve two-way binding of their specific data values. The
parent component now 'v-model's the correct value within the formData
object, and the custom component is (for all intents and purposes) blind
to it.

The base class functionality is divided into a few steps:
- custom editors can wrap their subcomponents into a top-level v-input
- the v-input

- the v-input is the interface for 2-way binding of the model value from
  the parent, and between the v-input and its subcomponents, by 'v-model'
  ing the 'internalValue'
- the 'useBaseInput' composable should be imported and used by a custom
  component. This provides the internalValue and the computed property
  getter and setter methods for updating 'internalValue'. For this reason
  the custom editor should provide two main functions as arguments to the
  composable:
  - a function to parse the 'v-model'ed values of the subcomponents from
    'internalValue', and
  - a function to determine 'internalValue' from the values of the subcomponents
  The composable also provides an object 'subValues' to which arbitrary
  keys can be added with which to 'v-model' the subcomponent values.
- The custom component should call defineEmits and pass the result to the
  composable, in order to establish a function that updates the parent on
  a local 'internalValue' change.

As a demonstration of how this can be implemented, the DateTimePickerEditor
and URIeditor components were updated according to this new framework.