Save entered form data to a graph
This commit adds functionality to take values from the formData
structure and convert it to quads that are saved to a new graph
store when a user selects the 'Save' button. The main addition
is the 'save_node' function in the formdata composable which
handles the quad creation and addition to a store (rdf.dataset).
In order to make the data store reactive such that the addition
of triples could be seen in real-time, a dummy ref was used to
force vue to react to changes. However, this stopgap should likely
be replaced by a smarter reactive data store implementation. Also,
it would make sense to have a single store for the graph data, both
for data loaded on application startup (e.g. anything existing in
some name space that is important for the user's context) and for
data saved upon form save. The reason is the referencing of blank
nodes, which would be consistent within the same store, but which
would cause errors across stores (is b14 in store A the same as b14
in form B? no).