Update form saving procedure to use reactive graph store
This introduces a reactive wrapper for an rdf.dataset inside the
graphdata composable. This allows actions like add and delete to
be picked up and reacted to by vue's reactivity system. This has
the consequence that the previous implementation of creating a new
separate dataset per form, implemented in the formdata composable,
could be removed. An important note is that the watch property can
be used on graphData in order to recalculate some dependent values