Changes to instanceselecteditor after updating graphstore
The process of generating the list of items to display in the
dropdown has now been moved to a function, which is called on
beforeMount as well as when the graphData changes (via watch
property). This solves the issue where an item that was saved
directly before from a dialog form didn't immediately appear
in its parent list.