Add info tooltip to items in InstanceSelectEditor
This addresses the issue (although not conclusively) of missing
information in a list item. The list item might only show the name
of a blank node, or of a named node, while this information might
not be enough for a user to recognize the item as the one they are
looking for. It is therefore useful to show additional information
about the item, which is now retrieved by querying the graphData for
all triples with the same subject as the item being listed