Fix recursive update loop that caused hanging
This (touch wood) fixes the longstanding issue where an infinite loop
caused the application to hang when the user selected 'Add new item'
from the autocomplete dropdown of an InstancesSelectEditor component.
The actual issue occurred in the NodeShapeEditor where the computed
property 'usedPropertyGroups' caused the component's onBeforeUpdate
hook to be called infinitely. This property was changes to a shallow
ref and it's value is now set once in the onMounted hook. This resulted
locally in the 'maximum updates exceeded' error disappearing. The next
test is to see if this persists when the site is built and served
remotely. In addition, small changes were made to an array ordering
function to rather order a copy than an array in-place (which 'order')
does).