Adds base functionlity for user to add a token
This adds the composables and components and further functions
for asking for a token to be entered when a user opts to submit
their edited data. The token is stored in sessionStorage and
subsequently made part of any future GET or POST request to
the configured service. Submit button is a simple icon in the
app header bar.