Allow immediate selection of an item that was generated from a recently saved form
This adds functionality to improve the UX of the instancesSelectEditor.
If a user adds a new item from this component, it opens a new form. The
expectation is that once the form is saved, this new item will immediately
be set as the selected item of the instancesSelectEditor. This functionality
was implemented by storing a reference to the specific instancesSelectEditor
that opened the form, and then also providing the identifying fields of the
saved form as arguments to the same context where the instancesSelectEditor
specifics are known. With this combination of information, the correct selected
item can be set in the associated instancesSelectEditor.