Introduce NodeShapeViewer and BlankNodeViewer
This is an effort to modularize the viewing functionality of shacl-vue
analogous to the form generation functionality. The previous RecordViewer
is replaced with a NodeShapeViewer (see NodeShapeEditor) which is a component
that provides a generic view of any named/blank node in the graph. This then
gives a view of all related triples (with all node types, blank/named/literal,
as objects), and specifically instantiates the BlankNodeViewer for blank nodes.
The BlankNodeViewer will show information recursively. In addition, the
getPrefLabel function was generalized and moved to utils to allow it to be used
across components. This was particularly useful for showing information in the
BlankNodeViewer in terms of the prefLabel and not the original text. To allow
this to be done on initial load, the application needed to be updated to get
specific data records from the integrated dumpthings service at startup. This is
because, e.g., an annotation needs to be locally known in order for it to be
accessible by the getPrefLabel function. This change was done by introducing
a new configuration variable 'service_fetch_before' which is an object with
which to specify the list of records that should be retrieved by ID and/or by class.