Refactor configuration handling
The goal here was to allow configuration to be used more seamlessly in
the app header component and main shaclvue component in order to drive
text values such as app name and urls and color themes and more. Before,
the appheader and appfooter were children of the main App.vue component,
along with the ShaclVue component. In order to prevent the convoluted
provide/inject setup that was necessary before, the header and footer were
moved into ShaclVue, becoming its children, and the main App.vue component
is now a single simple entrypoint for the ShaclVue component, without any
configuration or logic attached to it. New configuration options were
introduced:
- app name
- docs url
- source code url
- app theme: link color
- app theme: hover color
- app theme: active color
- app theme: left hand side panel color
- app theme: logo name
This also makes future additions to the theme easier and compartmenalized.