create two vite config files, one for library and one for application
The goal here is to be able to publish the library at dist/lib
to npm, while still being able to build the static files to
render a standalone site (from built files at dist/app). For the
library, the build.lib option was necessary to include, as well
as providing multiple entry points to allow multiple named imports
for apps that use shacl-vue.