Introduce property ordering by reverse inheritance
This introduces a default structure to the display of form fields of a specific class,
where groups of fields are ordered inversely according to the class in the hierarchy
that they inherit from. We call this the reverse hierarchical/inheritance display and
it replaces the ordering my sh:PropertyGroup. The latter is temporarily deactivated until
it will be reintroduced in a way that is compatible with the former, or such that it is
a configurable choice. Apart from the reverse inheritance grouping, no further property
ordering is introduced here. It remains a TODO to introduce a second level of ordering
based on e.g. required properties or some other annotation-based approach.

In addition, the default config is updated to show all properties by default (i.e. do
not hide unannotated/unrequired properties by default. The config also now includes
some default prefixes and class icons.