Helpers for getting default schemas and data
This is an interim solution in the form of shell scripts to fetch
datalad-concepts exports (shacl, owl) and to convert example data
to rdf, to allow using all of those as the default files for a
shacl-vue deployment. The ideal is to rather create a single base
schema in yaml, similar to the standard that we have followed in
other use cases e.g. trr379, inm7, etc, and to update the make files
that generate the schemas and data from there.