Introduce capability to send requests to multiple services
This feature involves a configuration update where the 'service_base_url'
option is changed from a single string to an array of objects of the form:
   {
      'url': 'https://metadata.inm7.de/simpleinput-unreleased/',
      'type': 'write'
   }
Other service-related configuration options remain unchanged. This is because
the new feature allows requests to be sent to multiple service with the
exact same API specifications and endpoint configurations, only the base url
changes. In addition, shacl-vue is configured to send GET requests to base urls
of all types, i.e. read and write (the latter implying the former), and to
send POST requests only to write-type base urls. A shacl-vue deployment
should be configured with one and only one write service, while it can be
configured with any number of read services. The new feature is made backwards
compatible with the previous manner of specifying the service url, via a single
string. In this case the base url is assumed to be of type write.

To achieve the above, two composables useData (specifically the fetchFromService
method) and useForm (specifically the submitFormData method) were updated.
They handle the two possibilities of base urls in the format of a single string
or an array. fetchFromService will now fetch from all urls specified via the
service_base_url config option, and submitFormData will only post to the single
write-type base url configured via service_base_url. If none or more than 1 base
urls of type write are found, errors are thrown and the submission is not
continued.