Update/fix property grouping and ordering in NodeShapeEditor
This is the second or third attempt to really *fix* this approach.
The previous method was found to be faulty when an exception showed
a property appearing in a group that it shouldn't appear in. It was
then updated to adhere to the approach defined by the following:

A property should be divided into the class group that either made the
latest change to it if it was inherited, or into the class that
introduced the property.

This necessitated several updates to achieve correct grouping and ordering.
Additionally, the updated method highlighted a non-ideal way in which
LinkML assings sh:order to property shapes of classes that are (also)
derived from mixins.

Therefore, while deploying this update to shacl-vue instances will
give these instances the updated ability to group and order properties
correctly, grouping and ordering might still appear faulty if the
underlying SHACL property shapes have faulty sh:order fields. For this
to work, these instances should also receive updated SHACL exports that
were generated by a patched version of LinkML that allows the exclusion
of automatic sh:order assignment.

Lastly, since the previous sh:PropertyGroup-based approach to ordering
is not used anymore, some remaining code relating to that feature was
removed, including the defaultPropertyGroup option in the config. Also,
the tab-layout was removed since it has never been used and has not
been kept in sync with the rest of the NodeShapeEditor development.