Introduce vue-virtual scroller, and more
This introduces vue-virtual-scroller as a dependency, and implements it

in the main ShaclVue component for displaying lists of records. This
components improves performance by not mounting every single item in
a list but rather mounting only those that are visible. The code uses
the dynamicscroller component rather than recyclescroller in order to
deal with variable heights of the NodeShapeViewer instances.

In addition, this commit also fixes the issue where 'no items' was shown
incorrectly for a short time after the skeleton loader disappeared and
before actual items started rendering.