New InstanceSelectEditor for improved performance and leaner data fetching
Most of the InstanceSelectEditor component was recreated or refactored. The
component has changed from using a vuetify autocomplete element at the base
to using a custom combination of a text field and menu. This was done in
order to accommodate the incorporation of a vue-virtual-scroller component
inside the new menu, which was complicated to achieve within the previous
autocomplete component.

Another major change was made to the order in which classes and their records
are calculated and fetched to support the functionality of the component.
Previously 'getAllRecordsFromService' was called from the setup script of the
component, after computing which classes are relevant for the given field.
Now, this fetch operation is only done once a user interacts with the component
which happens when they click on the fied in order to open the menu. In order
to show the actual value of the component once a form is opened, it is
necessary to fetch that individual record, so this part is done upfront (in
onBeforeMount).

The use of the debounce function from lodash-es (a new dependency) was
introduced to prevent short successive calls to the function that calculates
the list of items to show in a menu, in an effort to prevent unnecessary
lag effects. This call is done via a watch, which has also been updated to
watch a newly introduced ref on the rdfDS: rdfDS.data.graphChanged. This was
introduced because for some reason Vue was overeager in its reactivity tracking
of rdfDS.data.graph, so any computed ref or watch function that depended on
it was updated too often, which led to lag effects in some cases when the
function to calculate the list of items to render had to process many items.

Lastly, items in the menu are now filtered by preflabel.