Regression fix: 'no items' displaying prematurely when new type selected
This issue existed before, and was reportedly fixed in b239149316b135d1024bc222e6cdffee2ebd6c48.
However, the issue returned recently after several updates were made.

The latest approach is to watch the instanceItemsComp ref, which is a ref for
the list of items that should update whenever new data is fetched from the
backend service. It was first tried to set the classRecordsLoading value
to false in the rdfDS.data.graphChanged watcher, but this still showed the
'no items' for a second before then displaying the actual instanceItemsComp
list. So now a change was made to when the classRecordsLoading is set to false.
It is done in the instanceItemsComp watcher, but importantly not on any change,
since this ref changes each time a new data type is selected too. So  flag is
put in place to ignore the first trigger of the watcher and to run on the second.

However, it was soon after found that the watcher does not always trigger twice.
For data types that were selected previously and where the new selection will
cause the fetchFromService call to return 'skipped' for all requests made,
the classRecordsLoading also needed to be set to false, because otherwise
the ref watcher would not trigger and the loader would display infinitely. To
be able to know that all requests were skipped, the fetchFromService return
values had to be updated in the useData composable.

Overall, this is a house of cards. Let's see when the next regression happens.
For now I will leave it at this though.