Fix early display of 'no items'... again
The problem of the eternal loader happened again. This time it
was when the token was not entered for the backend service, and
two base urls were configured. And it only happened when selecting
a data type for the second time. So the requests were made, and
one failed (no token) and one skipped (because of a successful
previous fetch) but the existing logic did not account for this
combination of results. So this update not returns an array of
['success', 'failed', 'skipped'] possible options from the
fetchFromService routine. These are processed on return, and the
classRecordsLoading ref is set to false only if there were no
successful results (in the new 'status' field).