Add parseTTL method to ReactiveRdfDataset
This is in preparation for handling backend pagination requests.
The new pagination endpoint will return a json object with an 'items'
array containing TTL strings. The new function will parse a single
TTL string and add prefixes and quads to the RdfDataset. This is
opposed to the previous way in which an endpoint response was hanled,
where the response was a TTL string, and the shacl-tulip functionality
for fetching and directly processing the endpoint was used.