Introduce new configuration options to support git annex p2phttp
The new configuration mirrors the current state of the edu deployment,
and in addition adds the 'use_gitannex_p2phttp' option which needs to
be set to true of the deployment should allow file uploads. If true, then
an additional 'gitannex_p2phttp_config' option includes all required settings:
- deployment_url: e.g. https://hub.psychoinformatics.de, used for information
- info: arbitrary info to display to user
- info_url: a url with more information for the user
- base_url: the base url of annex p2phttp endpoints
- annex_uuid: the UUID of the git annex repository to which content will be uploaded
- client_uuid: the customizable UUID of the client
- id_prefix: the prefix to use when assigning PIDs to the Attachment records being created
  this prefix needs to be included in the 'prefixes' option of the same config

the configuration composable is updated to store these options in configVarsMain.

And the main app is temporarily updated to use this new config