New FileUploadEditor component
This component matches to a shape where the sh:class is 'inm7d:Attachment'.
This is not proposed as the ultimate way forward, but as a temporary showcase.
The component allows click+select or drag+drop of a file from the local filesystem
and then uses all the prior configration settings and entered credentials in order
to upload the file to a git annex p2phttp backend.

This has been tested locally only partially, because CORS settings do not allow
the localhost domain. So this could be error prone.