Introduce ShaclORClassEditor and update InstancesSelectEditor
This replaces the previous ShaclOREditor which did not function
as required to render an editor for a LinkML any_of slot, i.e.
a sh:or property. The new component has the same matching criterion
as the old component, i.e. sh:or should be in the property shape,
but in addition it matches only when all objects in the sh:or array
have the sh:class property. The component allows a user to select the
class from a dropdown, and then another dropdown appears from which the
use can actually select the relevant record (similar to the functioning
of the instancesselecteditor. The new component's modelvalue binds to
the IRI of the selected instance, just like the instancesselecteditor
does, and the selected class is not stored explicitly.

In addition, support for infinite scrolling using pagination requests
is introduced for both the ShaclORClassEditor and the InstancesSelectEditor.
Some common utility functions are factored out into the utils module.

The ShaclORClassEditor is fully functional except when its parent record is
being saved and the record subsequently submitted. The save operation saves
the selected IRI with an incorrect node kind, and the submission subsequently
fails. This requires a fix in shacl-tulip, see:
https://github.com/psychoinformatics-de/shacl-tulip/issues/12