Allow editing records from InstanceSelectEditor
This introduces a new config option, 'allow_edit_instances'. If set to true, this
will add an edit button for all instances in an 'InstancesSelectEditor' or a
'ShaclORClassEditor'. This edit button allows the user to edit the specific record
directly, without having to navigate to the record editor via the left-hand-side
panel. This edit button will be disabled if the record's class is included in
no_edit_classes. The config option works for both namedNode and blankNode records.