Introduce v1 of the 'W3CISO8601DateTimeEditor' component
This new editor component matches on property shapes where 'sh:datatype' == 'https://www.w3.org/TR/NOTE-datetime'.
It currently allows specifying dates in formats 'YYYY-MM-DD' (default), 'YYYY-MM', and 'YYYY':
- 'YYYY-MM-DD': uses a vuetify built-in date picker
- 'YYYY-MM': uses two autocomplete dropdowns, one each for year and month
- 'YYYY': uses a single autocomplete dropdown for year

A new 'W3CISO8601DateTimeEditor' option is also added to configuration. The 'yearStart' and 'yearEnd' options
can be used to define the starting and ending years that together define the range of options in the year-picker
for the 'YYYY-MM' and 'YYYY' formats. Respectively, they default to '1925' and '2077' if not specified. These
defaults are arbitrary at this point and should receive more attention.

Future addition of more formats are planned in order to support the complete W3CISO8601.