Introduce config-driven editor selection
The new 'config_selection' option allows the UI to use config-driven selection
of an editor component instead of the component matching procedure that 'shacl-vue'
uses by default. The object takes keys of a SHACL property shape as its keys in
CURIE format (e.g. 'sh:datatype', 'sh:path', or 'sh:nodeKind'), and the values
are objects themselves. These objects will have the to-be-matched CURIEs as keys,
and the corresponding value should be the exact name of the component that will be
selected. As an example, the following specifies that, if a SHACL property shape
is encountered where the 'sh:datatype' is equal to 'mydatetime:year', the
'W3CISO8601YearEditor' should be selected and rendered:

'editor_selection': {
    'sh:datatype': {
        'mydatetime:year': 'W3CISO8601YearEditor',
    }
}