Add ability to autogenerate record display label via string serialization
By default 'shacl-vue' uses the 'skos:prefLabel' of a record, if available,
as its display label. When not available, the new 'display_name_autogenerate'
option provides a means to autogenerate the display label of a record from a
string serialization of other properties of the same record. This option should
receive an object with class CURIEs as its keys and the values being string templates.
Placeholders in such a template shold be curly brackets containing the CURIE of a
property of the class that should be used instead of the placeholder.

When using the 'display_name_autogenerate' option, it is possible that not all parameters
in the template string exist as properties of an associated record for which a display
label is being autogenerated. For such cases, the missing parameter can be replaced with
a missing value placeholder string provided via the new 'display_name_autogenerate_placeholder'
option. This option allows for providing a 'default' placeholder to be used for all missing
values, or for providing a missing value placeholder per property.