Improve UX for infinite page fetching and progress indication
This became necessary due to bad UX when searching for specific records that have not been fetched at the time of searching. The goal was to have a seemless behavior that will continue to fetch new records (i.e. the next page from the server pagination endpoint) until a result is available, while integrating without issue with the existing behavior that fetches a new page when the last item in the list is scrolled to. The goal was to introduce these features for both the main page and for the InstancesSelectEditor.

The new state has the following behaviour:

- for both the main page and instance selector, show the progress values as GOT/TOTAL (i.e. locally available records vs total records on server), unless GOT >= TOTAL then only show GOT
- for both the main page and instance selector, show a visual linear progress bar also containing the progress values; for the main page, show this bar only on display name hover and when user does infinite search; for the instance selector, show it always
- always fetch first page of records when class records are viewed the first time (main page) or when a dropdown list is opened the first time (instance selector)
- fetch the next-in-line page of records when class records are viewed all following times or when a dropdown list is opened all following times
- when scrolling reaches the bottom of the class records or the bottom of the dropdown list of records, fetch the next-in-line page of records while showing a rotating circular progress icon at the bottom of the list
- filtering always happens against local list of records
- when a user adds filter-text, always show the linear progress bar
- when a user enters filter text that yields no local results, fetch the next-in-line page of records while showing a rotating circular progress icon
- for both the main page and instance selector, use a debouncing effect wrt the circular busy indicator which removes the show-hide stutter. Basically, if the page fetching happens many times in succession within a short start-stop time frame (which occurs when a user enters filter text that yields no initial results), the circular busy indicator will remain visible until the page fetching yields a result (i.e. the visibility flag has stopped toggling)

To support these changes, the useData Composable had to be updated to return pagination metadata for skipped requests.

TODO: these changes should also still be applied to the ShaclORClassEditor