Expand options for showing/hiding classes
Previously, the 'hide_classes' option was used to hide the provided list of class URIs from the left-hand-side panel listing all data types. A case emerged that motivated for expanding the show/hide functionality here. Specifically, it could be more efficient to explicitly include a set of class URIs to show, either as URIs or any classes with a given prefix, and then only to hide individual classes afterwards. For this reason, the complete new set of show/hide array options are:

- 'show_classes': show all of these classes
- 'show_classes_with_prefix': show all classes with these prefixes
- 'hide_classes': hide these classes
- 'hide_classes_with_prefix': hide all classes with these prefixes

If both 'show_classes' and 'show_classes_with_prefix' are empty arrays, all classes are shown, apart from those in hide_classes(_with_prefix).