Update `DLTYPES` namespace
Version 1 of `dltypes` has been released, and `dlthings/v1` is using it. This means that it will be pulled in by pretty much all instances.

This breaks the email editor component, which is looking for its type in this schema.

This change fixes that by updating the version.

However, I think the type (actually types) that this editor should act on better become configurable. With more versions coming, it will become unsustainable to prescribe a particular schema.