Continuous record fetching on constrained search
Before this update, a constrained search in both the main page and
'InstancesSelectEditor' would trigger a next-page-fetch function if
the filtered results were length zero. This happened, intentionally,
due to a complex interplay between the filtering function, the watcher
for the graph data being updated, and the trigger to fetch more records
when the last item in a list is viewed. This had two undesired aspects:
the complexity of the interplay between functions, and the fact that a
returned and filtered result of length > 0 would not trigger more fetch
operations. The latter is undesired because the user might be searching
for a record that is not returned and not available when the first
filtered result length > 0 is shown, and then no more fetching happens,
so the user is left waiting for nothing.

To improve this, the goal was firstly to update the behavior such that
if the user inputs search text, records will continue to be fetched
until all records are locally available. The second goal was to simplify
the complex interplay between functions. These goals were achieved by:

- changing 'filteredInstanceItemsComp' back to a computed property
- let 'filteredInstanceItemsComp' keep only the filtering code of
  the previous 'filterInstanceItems', not the part that conditionally
  fetches more records
- remove all calls to 'filterInstanceItems' from the code
- factor out the fetching of next pages into a 'async fetchNextPage()' function
- introduce a 'watchEffect' for the text field value that tracks changes to both
   the value as well as 'hasUnfetchedPages'; this allows us to continuously
   fetch more pages until the last page is fetched or until the text field value
   is empty
- don't call 'fetchNextPage' when 'isFetchingPage.value' is already true, thus
  preventing scrolling-fetching and filter-fetching from interfering with each
  other (or themselves)

These changes were applied to both the main page and 'InstancesSelectEditor'.

Two more changes were worked into this commit to improve the
'InstancesSelectEditor'. Firstly, a fix to a line of code that set the flag
that stores whether the 'InstancesSelectEditor' had been opened for the first
time. This wasn't working before, and caused the same funtion to run every
time a user clicks on the component, unnecessarily. Secondly, since it
became apparent that the dumpthings service will return all records of a
class and of all of its subclasses if a request is made to a specific class,
the 'InstancesSelectEditor' did not have to make requests to all of the
subclasses. These unnecessary steps were removed.