Fix ordering of properties for no-parent classes
The NodeShapeEditor is responsible for categorizing properties into
a set of hierarchical classes, for a form of any given class. E.g. if
the class 'Thing' has a property 'name' and the class 'SpecificThing'
is a subclass of 'Thing', then it too will have the property 'name'.
The NodeShapeEditor will divide this property into the display category
for one of its parent classes, based on a set of rules.

A bug was discovered for a case where the class for which a form is
rendered does not have a parent class. Here, the list of properties
to display were returned without being ordered by 'sh:order' first.

This commit fixes that oversight.