Tooltip updates and general styling improvements
Mainly: update tooltip functionality to improve UX for form field
descriptions.

For large screens:
- highlight the whole row on row hover
- underline label on row hover
- show cursor-pointer on row hover
- open tooltip when any location in the whole row is clicked,
  except on the actual form input field and associated buttons
- tooltip closes on outside click or on cursor exiting the row
  (i.e. no more hover)

For small screens:

- no hover effects
- label is underlined permanently
- open tooltip when any location in the whole row is clicked,
  except on the actual form input field and associated buttons
- tooltip has a cancel button that can be clicked to close it
- tooltip can also be closed on outside click

In addition, a uniform styling improvement was applied to a form
field row to improve the row hover display. This also solved the
issue where a required field validation error message was displayed
too far below the actual field, making it look like it was associated
with a different field.