Display actual response body for submission error
The reason the response body/text was unavailable for so long was because
the io module was using @rdfjs/fetch-lite which was wrapping the normal
fetch function and was handling its readable stream internally. By the
time the response.text() was accessed for error display, it was already
used.

Technically, the @rdfjs/fetch-lite library was not necessary for POST
requests, since the RDF was already serialized prior to providing it to
fetch-lite as the body, so the library's functionality for RDF handling
was unused.

Switching over to the standard fetch API solved these problems. This
necessitated a subsequent change to the useForm composable to handle to
changes return value, and then to the SubmitComp component in order to
display the response body correctlt. In addition, a copy butto was added
to allow the user to copy the error text to the clipboard.